pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cluster-experiment-runner"
version = "0.1.0"
description = "Submit, monitor, and cancel GPU experiments on SLURM clusters via SSH"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
cer = "cer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cer"]
